moved Board component and created actions and actionCreators for it. moved functions belonging to this component into actions,actioncreators,utils. Issue #6
